refactor: 通过 babel 解析实际引用文件
