+++ lib/mp-compiler/index.js
@@ -20,7 +20,8 @@ const {
   getSlots,
   htmlBeautify,
   getBabelrc,
-  getPageSrc
+  getPageSrc,
+  getImportsMap
 } = require('./util')
 
 let slotsHookAdded = false
@@ -248,12 +249,23 @@ function resolveSrc (originComponents, components, resolveFn, context) {
       resolveFn(context, originComponents[k], (err, realSrc) => {
         if (err) return reject(err)
         const com = covertCCVar(k)
-        const realComSrc = /\.js$/.test(realSrc)
-          ? path.join(path.dirname(realSrc), `${com}.vue`)
-          : realSrc
-        const comName = getCompNameBySrc(realComSrc)
-        components[com] = { src: comName, name: comName }
-        resolve()
+        const isScript = /\.js$/.test(realSrc)
+        if (isScript) {
+          const scriptContent = fs.readFileSync(realSrc).toString()
+          const result = babel.transform(scriptContent)
+          const { importsMap } = getImportsMap(result.metadata)
+          const source = importsMap[k]
+          resolveFn(path.dirname(realSrc), source, (err, realComSrc) => {
+            if (err) return reject(err)
+            const comName = getCompNameBySrc(realComSrc)
+            components[com] = { src: comName, name: comName }
+            resolve()
+          })
+        } else {
+          const comName = getCompNameBySrc(realSrc)
+          components[com] = { src: comName, name: comName }
+          resolve()
+        }
       })
     })
   }))
