+++ lib/mp-compiler/parse.js
@@ -2,23 +2,7 @@
 
 const generate = require('babel-generator').default
 const babelon = require('babelon')
-
-function getImportsMap (metadata) {
-  let { importsMap } = metadata
-  const { imports } = metadata.modules
-
-  if (!importsMap) {
-    importsMap = {}
-    imports.forEach(m => {
-      m.specifiers.forEach(v => {
-        importsMap[v.local] = m.source
-      })
-    })
-    metadata.importsMap = importsMap
-  }
-
-  return metadata
-}
+const { getImportsMap } = require('./util')
 
 // 解析 config
 const traverseConfigVisitor = {
