+++ lib/mp-compiler/util.js
@@ -119,6 +119,23 @@ function getPathPrefix (src) {
   return `${'../'.repeat(length)}`
 }
 
+function getImportsMap (metadata) {
+  let { importsMap } = metadata
+  const { imports } = metadata.modules
+
+  if (!importsMap) {
+    importsMap = {}
+    imports.forEach(m => {
+      m.specifiers.forEach(v => {
+        importsMap[v.local] = m.source
+      })
+    })
+    metadata.importsMap = importsMap
+  }
+
+  return metadata
+}
+
 const defaultStylePart = {
   type: 'style',
   content: '\n',
@@ -146,5 +163,6 @@ module.exports = {
   htmlBeautify,
   getBabelrc,
   getPathPrefix,
-  getPageSrc
+  getPageSrc,
+  getImportsMap
 }
